add tcl location
