+++ src/nx/__init__.py
@@ -1,4 +1,4 @@
 
 from .dxf_export import DxfExporter
-from .msgbox import error, warn, info
+from .dialog import error, warn, info
 from .utils import *
